Admin Panel

+++ pages/api/order/index.tsx
@@ -16,35 +16,45 @@ export default withIronSessionApiRoute(async function handler(
   const userRepository = connection.getRepository(UserEntity);
   const orderRepository = connection.getRepository(OrderEntity);
 
-  const basketItems = await connection
-    .getRepository(BasketItemEntity)
-    .createQueryBuilder('BasketItem')
-    .leftJoinAndSelect('BasketItem.item', 'Item')
-    .leftJoinAndSelect('BasketItem.user', 'User')
-    .where('BasketItem.user = :id', { id: req.session.user })
-    .getMany();
+  if (req.method == 'POST') {
+    const basketItems = await connection.getRepository(BasketItemEntity).find({
+      relations: {
+        item: true,
+        user: true,
+      },
+      where: {
+        id: req.session.user,
+      },
+    });
 
-  const { tel, email } = req.body;
-  const user = await userRepository.save({
-    id: req.session.user,
-    tel,
-    email,
-  });
+    const { tel, email } = req.body;
+    const user = await userRepository.save({
+      id: req.session.user,
+      tel,
+      email,
+    });
 
-  const order = await orderRepository.save({
-    date: new Date(),
-    user: user.id,
-    orderedItems: basketItems.map((basketItem: BasketItem) => {
-      return {
-        item: basketItem.item,
-        quantity: basketItem.quantity,
-      };
-    }),
-  });
+    const order = await orderRepository.save({
+      date: new Date(),
+      user: user.id,
+      orderedItems: basketItems.map((basketItem: BasketItem) => {
+        return {
+          item: basketItem.item,
+          quantity: basketItem.quantity,
+        };
+      }),
+    });
 
-  const clearBasketItems = await basketItemRepository.delete({
-    user: req.session.user,
-  });
-  res.status(201).json(clearBasketItems);
+    const clearBasketItems = await basketItemRepository.delete({
+      user: req.session.user,
+    });
+    res.status(201).json(clearBasketItems);
+  } else if (req.method == 'PUT') {
+    const order = await orderRepository.save({
+      id: req.body.id,
+      status: req.body.status,
+    });
+    res.status(201).json(order);
+  }
 },
 SESSION_OPTIONS);
